refactor(console): merge utilities folder
