+++ packages/console/src/utilities/regex.ts
